Added 7 new default hammers
Added Tin, Copper, Silver, Lead, Platinum, Nickel and Aluminum Hammers.

While the hammer items are initialized at the pre-init phase, the actual
tool enum to be used in those items is assigned at the post-init phase.
They are separated in order to properly detect the tool enum in use
without adding "after" dependencies to this mod.

The enum is first retrieved from the recipe list with a pickaxe recipe.
If the enum does not exist, a call to EnumToolMaterial.valueOf would be
used to retrieve a fallback. If the enum still does not exist, a final
hardcode fallback specified in the AliquoMaterial class would be used.

A class AliquoRegistry is also added to handle textures.